noteit_backend/requirements.txt
Django>=5.0
mysqlclient
python-dotenv
djangorestframework
django-cors-headers